Out with the old, in with the new.
Overall cleanup to style guidelines (variable naming and such). Introduce World, Camera, and Box. Rip out the hooks to Map which will soon be transferred to World. Major work on drawing remains (see Box.cs). Making static map object drawing work correctly is a priority. Functionality to pan across the (non-server/file generated) world is included with the arrow keys, zooming with pgup/pgdown, and rotation with left and right shift keys.

+++ AngryTanks.Client/StaticSprite.cs
@@ -8,7 +8,7 @@ namespace AngryTanks.Client
 {
     public class StaticSprite : Sprite
     {
-        public StaticSprite(Texture2D texture, Vector2 position, Vector2 size, Single rotation)
+        public StaticSprite(Texture2D texture, Vector2 position, Vector2 size, Double rotation)
             : base(texture, position, size, rotation)
         {
         }
